Eliminated use CON_time. Replaced with use ModTimeConvert, and adding a local tStart variable which is set by init_axes.
Introduced DoUpdateB0 logical instead of using a negative sign for DtUpdateB0
This way DtUpdateB0 can be set in advance for a later session.
Simplified logic in set_axes.

In fact this has probably fixed a bug: time accurate following a steady state
run would have made DtUpdateB0 negative in the first session so the
time accurate session would not behave as expected.


Former-commit-id: 7b652b3f0c42978c66a9f46c74f5ca751bf0a967