Removed the swmfclean target (was same as distclean).
Remved the clobber target (same as distclean and rm -rf run).

The original distclean target is renamed to distclean_comp.

The recently created veryclean target is renamed to distclean.
(removes the SC/BATSRUS/src and IH/BATSRUS/src directories and
all files not in the distribution).
This is the target used by SetSWMF.pl -uninstall and make dist.

The SCBATSRUS and IHBATSRUS targets are shortened and cleaned up.
In the help target several spelling errors are fixed.


Former-commit-id: 44a443f85fbfc98cb4fb856935fbec0fd3738311