Bug fix: mutiple radii for 'flx' variables did not work. Now fixed.
'dstflx' and 'dstdivb' log variables are available to calculate
the surface integral of B1z and the volume integral of z*div B/r**3,
respectively. The surface integral is taken at the radii listed for
'flx' variables, while the volume integral of 'dst' and 'dstdivb' has
the inner boundary at rCurrents.


Former-commit-id: 8c415980ff71328914756f0b8e3b417b6c69faa5