GM/BATSRUS/src/MH_set_parameters.f90: Bug fix: call init_mod_raytrace all the time so now dynamic allocation works for GM-IM or GM-RB coupling.
Condition if(UseIM) was wrong for several reasons: UseIM is not set at
this point, arrays for ray integral must be allocated even if IM is not
used.

This fix is a waste of memory if ray tracing is not used, but it is robust
and simple. The ray tracing can always be configured out to save memory.


Former-commit-id: 0019462836502aa39ddd5c2f6cda389f6a0e29f7