Fixed the problem that satellite file names did not contain time step, so subsequent restarts would overwrite the files. The satellite file name was
satellite_NN_NAME.sat

where NN is the satellite file index (1..nSatellite) and NAME
is the name of the trajectory file without extension.
This is now changed to

sat_NAME_nNoSTEP.sat

where NoSTEP is the n_step variable in BATSRUS. The satellite index
has been intentionally dropped so that the order of the satellite files
will make no difference in the name.


Former-commit-id: a758e7200549e11a2bd7ff5cbd9347349dde8bbf