Only covariant files are modified.
1. Before the present changes the face area vector (three components) had not
been stored, only its magnitude had been. The face area vector used in the
flux/source computation recovered in terms of the stored magnitude and its
direction was taken along the line, connecting the centers of the neighboring
cells.

Now all three components of the face area vector are stored. This way requires
more memory, but allows us to use a larger variety of covariant grids, such
that the face area vector may be not oriented along the line, connecting the
cell centers.

2. For center-based covariant grids it is desirable, and for vertex-based grids
it is mandatory to express the face area vectors and the volume of cells in
terms of the node coordinates. This procedure is implemented now and old
procedures, which work only for some simplest geometries, are deleted.

 Modified Files:
 	src/ModCovariant.f90 src/calc_flux_aw_covar.f90
 	src/calc_flux_aw_modif_covar.f90 src/calc_flux_linde_covar.f90
 	src/calc_flux_linde_covar_mars.f90
 	src/calc_flux_rusanov_covar.f90 src/calc_sources_covar.f90
 	src/covariant.f90 src/set_block_geometry.f90


Former-commit-id: 5043f64e1df80874f9c82119be5cc92b5e826cc1