Eliminated the use of cTiny and division by GradNormal2 which caused inaccurate results for smooth problems.
Commented out the limiting of the transverse gradients because it does
not seem to be correct (makes the scheme first order accurate if
the normal gradient is zero).

Use the MC-beta limiter for the TVD res. change algorithm
if the MC-beta limiter is used elsewhere.

Added the 'mc3' limiter which should provide a 3rd order finite difference
scheme for uniform grid and pointwise source terms.


Former-commit-id: a8fe1d08dd0bf699b4c6d00197c2aa11378d64e3