This message was intended for the previous change.
Modified functionality test to use 720 blocks instead of 400 / proc
so that it can do the full physics based AMR on 1 processor.
This way results will not depend on rounding errors, and will
be the same as for the 2 and more processor runs.
This should fix the problems on grendel and nyx.


Former-commit-id: dc4475268ac84dd3197dc28a4617cf36a64c69e5