The bug is fixed in the ModUserCrash for the "cylidrical" grid. The number of added lines: 2
In update_states routine the face areas and control volumes are multiplied by
some geomeetric factors, to account for the effect of the cylindrical geometry.
However, in determining the time step allowed for in the explicit scheme the
corresponding geometric factor was not introduced as required. The latter is wrong
and is assumed to be evidently wrong, because within the control volume approach,
the condition for the scheme stability is also controlled by face-to-volume ratio.

The fix: in update_state routine along with the multiplication of fluxes and control
volumes by some geometric factors, the contribution to the CFL criterion, VDtFace_Y
is multiplied by the ratio of the geomtric factors, for the face Y and the control
volume.

In the test run the first time step is reduced by ~10% as the result of the algorithm
modification. The latter means that the CFL, although claimed to be 0.8 was aclually
closer to 0.9 before. In  high resolution runs, presumably, the local CFL near pole
might exceed 1.1-1.2.

Modified Files:
 	GM/BATSRUS/Param/CRASH/TestOutput/log_levelset.log
 	GM/BATSRUS/srcUser/ModUserCrash.f90


Former-commit-id: 74e2aed38f9e75a491e2bc8308f305a64c0bcdd1