The ability to save traced MHD values, and read them in in future simulations, has been added. This feature allows us to do RBE simulations multiple times, while only needing to run the MHD code one time. Saving a tremendous amount of CPU time and effort.
The methods for doing the reading and writing is in ModPrerunField.

Options for writing the traced field, and using the previously traced
field, are set in set_parameters

rbe_main, rbe_subs, and RB_wrapper are modified to include calls to
read and write routines.

Makefile is modified to include ModPrerunField in object list.



 Modified Files:
 	src/Makefile src/ModRbe.f90 src/rbe_main.f90 src/rbe_subs.f90
 	src/set_parameters.f90 srcInterface/RB_wrapper.f90
 Added Files:
 	src/ModPrerunField.f90


Former-commit-id: b301fedc56ff5664e5fa95f88e8fc9cff4fa4d4f