Added a new method for setting the initial condition in CRCM. Below is a summary of the required changes:
The #INITIALF2 command in the param.in has now changed. There are three
options. Use and empty initial condition, use GM to set the initial
condition, or use observations to set the initial conditions. Old
PARAM files need to be updated now to work with the method. The test6
PARAM files are updated accordingly.

The option to set the initial condition from data works with CRCM configured
with H+, O+ and e, however it does not currently work if He+ ions are included.

The data for the new initial condition is stored in the newly created
data directory.





 Modified Files:
 	IM/CRCM/Makefile IM/CRCM/src/ModInitialize.f90
 	IM/CRCM/src/crcm.f90 IM/CRCM/src/set_parameters.f90
 	Param/PARAM.in.test.GMIEIMCRCM
 	Param/PARAM.in.test.restart.GMIEIMCRCM
 Added Files:
 	IM/CRCM/data/quiet_e.fin IM/CRCM/data/quiet_h.fin
 	IM/CRCM/data/quiet_o.fin


Former-commit-id: bd7d4f38b838d1dccb2460bbe9b059fac3a1931b