he updated version of the Meng Jin's SWMF_GLSETUP.pro script. Here,
read_fits.py
GLSETUPAlg.py

shape a  fully python part of the script. At the same time they are agreed to
be used as the "back end" ("server side") of the EEGGL web application to
substitute Java scripts: may be, without reading fits file with python, but
with keeping a new capabilities for smoothing and remapping the magnetogram,
which both bring quite new and useful opportunity for pre-processing and
visualizing, the magnetogram.

GLSETUP.py

is a master script, in which the IDL leftover of the original
Meng's script,

GLSETUP1.pro
GLSETUP2.pro

are called as a child subprocess. These fiels allow to the SWMF user to
fully keep for the SWMF user the preserved capability of SWMF_GLSETUP.pro
for IDL vizualization of the results. As far as it concerns EEGGL, these
three files can be considered as the emulator of the "front end"
("browser side") of the web application, to be implemented in JAVA-script
or HTML language.

List of things to do:
Implement an algorithm for UsePIL=True. Currently this piece of the code is
not even compilable.


Former-commit-id: 0061f2e8249adae2f81c78af3b1c04c848165397