Some chages by Adam, plus the following fixes: Obsolete Mneutralsmp replaced with mProtonMass set by #OHNEUTRALS. Removed obsolete #NEUTRALS command.
Initialized Ustar_I, Rate_I, UstarPu3_I, RatePu3_I so that first element
corresponding to the total fluid (which is NOT needed) is set.

Initialized UStarM_I and UStarMPu3_I to 1.0 instead of 0.0 to
avoid division by zero. This was already done in ModOuterHelio.f90

Fixed the size of the iVarPointImpl_I array (20 --> 30).


Former-commit-id: e990ea6c7845f0e4d9cc82fa8f900f8eec2ba760